safari-534.53-branch/LayoutTests/http/tests/media/video-referer-expected.txt

EVENT(canplaythrough)
Tests that the media player will send the relevant referer when requesting the media file.
